+++ inlang/source-code/paraglide/paraglide-js/src/cli/commands/init.ts
@@ -26,10 +26,10 @@ export const initCommand = new Command()
 		await checkIfUncommittedChanges()
 		await checkIfPackageJsonExists()
 		const projectPath = await initializeInlangProject()
-		const namespace = await promptForNamespace()
-		await addCompileStepToPackageJSON({ projectPath, namespace })
-		await adjustTsConfigIfNecessary()
 		await addParaglideJsToDependencies()
+		await addCompileStepToPackageJSON({ projectPath })
+		await maybeChangeTsConfigModuleResolution()
+		await maybeChangeTsConfigAllowJs()
 		await maybeAddVsCodeExtension({ projectPath })
 
 		consola.box(
@@ -99,27 +99,6 @@ export const addParaglideJsToDependencies = async () => {
 	pkg.dependencies["@inlang/paraglide-js"] = version
 	await fs.writeFile("./package.json", stringify(pkg))
 	consola.success("Added @inlang/paraglide-js to the dependencies in package.json.")
-}
-
-export const promptForNamespace = async (): Promise<string> => {
-	let assumedName = process.cwd().split("/").pop()
-	try {
-		assumedName = JSON.parse(await fs.readFile("./package.json", { encoding: "utf-8" })).name
-	} catch {
-		// nothing
-	}
-
-	consola.info(`You need to select a name for the project.
-
-The name is used to create an importable 'namespace' to distinguish between multiple projects in the same repository. For example, the name \`frontend\` will create the following import statement:
-	
-\`import * as m from "@inlang/paraglide-js/frontend/messages"\``)
-
-	const namespace = await prompt(`What should be the name of the project?`, {
-		type: "text",
-		initial: assumedName,
-	})
-	return namespace.trim()
 }
 
 export const findExistingInlangProjectPath = async (): Promise<string | undefined> => {
@@ -246,10 +225,7 @@ export const checkIfUncommittedChanges = async () => {
 	}
 }
 
-export const addCompileStepToPackageJSON = async (args: {
-	projectPath: string
-	namespace: string
-}) => {
+export const addCompileStepToPackageJSON = async (args: { projectPath: string }) => {
 	const file = await fs.readFile("./package.json", { encoding: "utf-8" })
 	const stringify = detectJsonFormatting(file)
 	const pkg = JSON.parse(file)
@@ -257,15 +233,15 @@ export const addCompileStepToPackageJSON = async (args: {
 		if (pkg.scripts === undefined) {
 			pkg.scripts = {}
 		}
-		pkg.scripts.build = `paraglide-js compile --project ${args.projectPath} --namespace ${args.namespace}`
+		pkg.scripts.build = `paraglide-js compile --project ${args.projectPath}`
 	} else if (pkg?.scripts?.build.includes("paraglide-js compile") === false) {
-		pkg.scripts.build = `paraglide-js compile --project ${args.projectPath} --namespace ${args.namespace} && ${pkg.scripts.build}`
+		pkg.scripts.build = `paraglide-js compile --project ${args.projectPath} && ${pkg.scripts.build}`
 	} else {
 		consola.warn(`The "build" script in the \`package.json\` already contains a "paraglide-js compile" command.
 
 Please add the following command to your build script manually:
 
-\`paraglide-js compile --project ${args.projectPath} --namespace ${args.namespace}\``)
+\`paraglide-js compile --project ${args.projectPath}`)
 		const response = await consola.prompt(
 			"Have you added the compile command to your build script?",
 			{
@@ -292,7 +268,7 @@ Please add the following command to your build script manually:
  * Otherwise, types defined in `package.exports` are not resolved by TypeScript. Leading to type
  * errors with Paraglide-JS.
  */
-export const adjustTsConfigIfNecessary = async () => {
+export const maybeChangeTsConfigModuleResolution = async () => {
 	if (fsSync.existsSync("./tsconfig.json") === false) {
 		return
 	}
@@ -339,7 +315,7 @@ export const adjustTsConfigIfNecessary = async () => {
 			`Did you set the \`compilerOptions.moduleResolution\` to "Bundler"?`,
 			{
 				type: "confirm",
-				initial: false,
+				initial: true,
 			}
 		)
 		if (response === false) {
@@ -358,6 +334,55 @@ export const adjustTsConfigIfNecessary = async () => {
 		} else {
 			consola.error(
 				"The compiler options have not been adjusted. Please set the `compilerOptions.moduleResolution` to `Bundler`."
+			)
+		}
+	}
+}
+
+/**
+ * Paraligde JS compiles to JS with JSDoc comments. TypeScript doesn't allow JS files by default.
+ */
+export const maybeChangeTsConfigAllowJs = async () => {
+	if (fsSync.existsSync("./tsconfig.json") === false) {
+		return
+	}
+	const file = await fs.readFile("./tsconfig.json", { encoding: "utf-8" })
+	// tsconfig allows comments ... FML
+	const tsconfig = JSON5.parse(file)
+
+	if (tsconfig.compilerOptions?.allowJs === true) {
+		// all clear, allowJs is already set to true
+		return
+	}
+
+	consola.info(
+		`You need to set the \`compilerOptions.allowJs\` to \`true\` in the \`tsconfig.json\` file:
+
+\`{
+  "compilerOptions": {
+    "allowJs": true
+  }
+}\``
+	)
+	let isValid = false
+	while (isValid === false) {
+		const response = await prompt(`Did you set the \`compilerOptions.allowJs\` to \`true\`?`, {
+			type: "confirm",
+			initial: true,
+		})
+		if (response === false) {
+			return consola.warn(
+				"Continuing without adjusting the tsconfig.json. This may lead to type errors."
+			)
+		}
+		const file = await fs.readFile("./tsconfig.json", { encoding: "utf-8" })
+		const tsconfig = JSON5.parse(file)
+		if (tsconfig?.compilerOptions?.allowJs === true) {
+			isValid = true
+			return
+		} else {
+			consola.error(
+				"The compiler options have not been adjusted. Please set the `compilerOptions.allowJs` to `true`."
 			)
 		}
 	}