+++ inlang/source-code/website/src/pages/index/custom_section/Hero.tsx
@@ -4,7 +4,7 @@ import { Chip } from "#src/interface/components/Chip.jsx"
 import { colorForTypeOf, typeOfIdToTitle } from "#src/pages/m/utilities.js"
 import { Button } from "../components/Button.jsx"
 import Link from "#src/renderer/Link.jsx"
-import * as m from "@inlang/paraglide-js/website/messages"
+import * as m from "#src/paraglide/messages.js"
 
 const featuredArray = [
 	"library.inlang.paraglideJs",